refactor: move city routes to avoid conflicts
Relocate city pages to '/city/[city]/' to prevent routing conflicts.
